Add a pyproject.toml file
pyproject.toml is the current standard way of defining metadata for
a Python package. In this commit I'm adding a pyproject.toml that stores
the typical core metadata (package name, author, etc.). It also records
the fact that the main() function from the last commit should be
the entry point when pip/pipx/etc. generates a wrapper script (a console
script). And it defines hatch as the build system for building
distributable packages, which will support making pypipe available on
PyPI.

I chose hatch somewhat arbitrarily, but I think it has a few good things
going for it:

- it's designed for Python projects that have a single package or module
  like pypipe does (just pypipe.py, or if the code ever gets split up,
  it could be moved into multiple files under the pypipe/ directory)
- it's relatively simple and easy to configure, and directly supports
  pyproject.toml, unlike something like setuptools
- it supports plugins that can determine the version from the Git tag,
  in case pypipe's maintainers ever want to do that

That being said, it wouldn't be hard to switch to another build backend
if desired.

As discussed on the pull request
  https://github.com/bugen/pypipe/pull/8#discussion_r1379798279
the pyproject.toml file defines three console scripts, all of which
invoke the same backend function. This supports installation with
`pipx install` or `pip install --user`, for people who would like to
install the package that way.