Add a release workflow
In this commit I've added a configuration file for a Github Actions
workflow that will build pypipe into a package, test it, and upload it
to PyPI. I based this on workflows I've used in my own projects, most
notably setuptools-pyproject-migration, but it's somewhat simplified
because it doesn't include documentation and has only rudimentary logic
to run tests. I'm working under the assumption that this can be improved
later. (e.g. potentially by using tox, pre-commit, etc.)

The testing logic is adapted from ci.yml, but modified to make sure that
it runs tests using the same built package that is going to be uploaded
to PyPI if the tests pass.

The workflow uses trusted publishing to upload the package, which is
relatively secure in the sense that it allows only this specific Github
Actions workflow to do the upload without any passwords or tokens being
involved. This ensures that packages on PyPI come from a known source
and are verifiable.