Checkout code and setup Python in pre-release testing
The release workflow was failing because I forgot to add a step that
actually checks out the code before running pre-release testing, so
there were no test files there to run. This change fixes that: it checks
out the repository before trying to run pytest, so now pytest can find
its configuration and the test files and it should be able to actually
run.

I also added a setup-python step to prepare the Python interpreter. This
is not strictly necessary, I think, but it's a pretty standard thing to
have in other projects, and it will be useful if we try to test on
specific Python versions in the future.