Ignore SIGINT when pager is active.
When using less or more as a pager, interrupting with Ctrl-C may cause display issues (terminal input becomes invisible). While the exact cause is unclear, ignoring SIGINT on the pypipe side can mitigate the problem. This issue occurs with less and more but not with most. Exiting the pager with 'q' does not trigger such problems. Using bat as a pager results in a similar issue; however, applying this fix does not resolve it. Like less and more, exiting bat with 'q' avoids any issues.
Furthermore, in line with this change, the default pager command has been updated from 'less -R -F -K' to 'less -R -F'.