Adjust the handling of Ctrl-C (#25)
- Modified to handle SIGINT depending on the situation.

When pypipe is mistakenly executed without using a pipe,
it continues to wait for standard input and does not exit.
In such cases, you may want to terminate it with Ctrl-C (SIGINT).
However, when the pager is active, if SIGINT is accepted
while pypipe is waiting for the termination of the pager in the exit
process (on_exit), it may lead to corrupted terminal display
(entering non-echo mode). Therefore, it has been adjusted to accept
SIGINT until just before the processing reaches on_exit,
and to ignore it if it hasn't reached on_exit yet.

- Modified to launch the pager in a separate process group.

By adding the option start_new_session=True to subprocess.Popen,
the pager is now launched in a separate process group from pypipe.
In the case of a pager that accepts SIGINT (e.g., bat), even if pypipe
ignores SIGINT, the pager may still terminate with SIGINT, causing
potential issues with the terminal state (entering non-echo mode).
Although the reason for this behavior is not fully understood,
separating the process groups prevents the pager from terminating
with Ctrl-C. So far, no issues have arisen due to this change.
If any problems are discovered, this modification may be rolled back.